# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TTL-Legacy Backend API
  description: >
    Backend API for TTL-Legacy, a decentralized "Dead Man's Switch" built on Stellar/Soroban.
    This API provides endpoints for vault management, reminder preferences, notifications, and administrative operations.
  version: 1.0.0
  contact:
    name: TTL-Legacy Team
    url: https://github.com/TTL-Legacy/TTL-Legacy
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.ttl-legacy.app
    description: Production server
  - url: https://testnet-api.ttl-legacy.app
    description: Testnet server
...
paths:
...
  /api/vaults/{vault_id}/reminder-preferences:
    parameters:
      - name: vault_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Unique identifier of the vault
    get:
      summary: Get reminder preferences for a vault
      description: Retrieve the current reminder notification preferences for a specific vault
      operationId: getPreferences
      tags:
        - Reminder Preferences
      responses:
        '200':
          description: Reminder preferences retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderPreferences'
        '404':
          description: Vault not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Set reminder preferences for a vault
      description: >
        Update or create reminder notification preferences for a vault.
        Supports idempotent requests via the `idempotency-key` header.
      operationId: setPreferences
      tags:
        - Reminder Preferences
      parameters:
        - name: idempotency-key
          in: header
          required: false
          schema:
            type: string
            format: uuid
          description: >
            UUID to ensure idempotent requests. If provided, the same response will be returned
            for identical requests within 24 hours.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPreferencesRequest'
            examples:
              emailOnly:
                summary: Email-only reminders
                value:
                  channels:
                    - email
                  hours_before_expiry: 24
                  frequency: daily
              multiChannel:
                summary: Multiple notification channels
                value:
                  channels:
                    - email
                    - sms
                    - push
                  hours_before_expiry: 48
                  frequency: weekly
      responses:
        '200':
          description: Preferences set successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderPreferences'
        '400':
          description: Invalid request (empty channels or invalid hours_before_expiry)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                emptyChannels:
                  value:
                    error: "channels must not be empty"
                invalidHours:
                  value:
                    error: "hours_before_expiry must be > 0"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/vaults/{vault_id}/simulate-release:
    get:
      summary: Simulate vault release scenarios
      description: >
        Projects when a vault will release to its beneficiary under one or more
        check-in behaviour scenarios. Useful for owners who want to understand
        the impact of their check-in habits on release timing.
      operationId: simulateVaultRelease
      tags:
        - Release Simulator
      parameters:
        - name: vault_id
          in: path
          required: true
          description: The ID of the vault to simulate
          schema:
            type: string
          example: vault-abc-123
        - name: scenarios
          in: query
          required: false
          description: >
            Comma-separated list of scenarios to run. Defaults to all three when
            omitted. Valid values: `no_check_ins`, `consistent_check_ins`,
            `missed_check_in_dates`.
          schema:
            type: string
            example: "no_check_ins,missed_check_in_dates"
        - name: missed_count
          in: query
          required: false
          description: >
            For the `missed_check_in_dates` scenario: number of consecutive
            check-ins the owner misses before stopping entirely. Defaults to 1.
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 2
      responses:
        '200':
          description: Simulation results for the requested scenarios
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulateReleaseResponse'
              example:
                vault_id: "vault-abc-123"
                current_ttl_remaining: 86400
                check_in_interval: 2592000
                last_check_in: "2026-06-29T10:00:00Z"
                simulated_at: "2026-06-30T06:39:00Z"
                scenarios:
                  - scenario: "no_check_ins"
                    description: "Owner performs no further check-ins. Vault releases when the current TTL expires."
                    projected_release_at: "2026-07-01T06:39:00Z"
                    seconds_until_release: 86400
                    confidence: "high"
                    notes: "Current TTL remaining: 86400 seconds (1.0 days)."
                  - scenario: "consistent_check_ins"
                    description: "Owner checks in consistently at the configured interval. Vault does not release."
                    projected_release_at: "2126-06-30T06:39:00Z"
                    seconds_until_release: -1
                    confidence: "high"
                    notes: "With a check-in interval of 2592000 seconds (30.0 days), consistent check-ins prevent vault release indefinitely."
                  - scenario: "missed_check_in_dates"
                    description: "Owner misses 1 consecutive check-in(s), then stops entirely."
                    projected_release_at: "2026-08-01T06:39:00Z"
                    seconds_until_release: 2678400
                    confidence: "medium"
                    notes: "Each missed check-in adds 2592000 seconds (30.0 days) to the release window. 1 missed check-in(s) → 2592000 additional seconds."
        '404':
          description: Vault not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Invalid scenario name(s) specified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Legal Document Anchoring (Issue #1339) ─────────────────────────────────
  /api/vaults/{vault_id}/document-anchors:
    post:
      summary: Anchor a legal document hash
      description: |
        Records the SHA-256 hash of a signed legal document as an immutable
        on-chain anchor tied to the vault.

        **Raw document bytes are never sent to this endpoint** — only the
        64-character hex-encoded SHA-256 hash is transmitted.

        **Legal disclaimer**: this does not constitute legal execution of the
        underlying document. It provides a cryptographic timestamp proof of
        the document's existence. Consult a qualified legal professional for
        estate-planning advice.
      operationId: createDocumentAnchor
      tags:
        - Legal Document Anchoring
      parameters:
        - name: vault_id
          in: path
          required: true
          schema:
            type: string
          description: Numeric vault ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnchorDocumentRequest'
            examples:
              will:
                summary: Anchor a will
                value:
                  owner_address: "GABCDE..."
                  doc_hash_hex: "a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2"
                  doc_type: "will"
                  storage_ref: "ipfs://QmExampleCid"
      responses:
        '201':
          description: Document anchored successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentAnchorRecord'
        '422':
          description: Invalid input (e.g. hash not 64 hex chars)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: List document anchors for a vault
      description: |
        Returns all document anchor records for the vault, including
        soft-removed ones. Filter client-side on `removed = false` to display
        only active anchors.
      operationId: listDocumentAnchors
      tags:
        - Legal Document Anchoring
      parameters:
        - name: vault_id
          in: path
          required: true
          schema:
            type: string
          description: Numeric vault ID
      responses:
        '200':
          description: List of document anchors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentAnchorRecord'
        '422':
          description: Invalid vault_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/vaults/{vault_id}/document-anchors/{doc_id}:
    delete:
      summary: Soft-remove a document anchor
      description: |
        Marks a document anchor as removed (`removed = true`). The hash record
        is preserved permanently for audit purposes — on-chain history cannot
        be mutated.
      operationId: deleteDocumentAnchor
      tags:
        - Legal Document Anchoring
      parameters:
        - name: vault_id
          in: path
          required: true
          schema:
            type: string
          description: Numeric vault ID
        - name: doc_id
          in: path
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Sequential document ID within the vault
      responses:
        '204':
          description: Anchor soft-removed successfully (no body)
        '404':
          description: Anchor not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        status_code:
          type: integer
          description: HTTP status code
      required:
        - error
...
    Channel:
      type: string
      enum: [email, sms, push]
      description: Notification delivery channel
    Frequency:
      type: string
      enum: [once, daily, weekly, hourly, monthly]
      description: Reminder frequency
    ReminderPreferences:
      type: object
      properties:
        vault_id:
          type: string
          format: uuid
          description: Vault ID
        channels:
          type: array
          items:
            $ref: '#/components/schemas/Channel'
          description: Notification channels enabled
        hours_before_expiry:
          type: integer
          minimum: 1
          description: Hours before expiry to send reminder
        frequency:
          $ref: '#/components/schemas/Frequency'
      required:
        - vault_id
        - channels
        - hours_before_expiry
        - frequency
    SetPreferencesRequest:
      type: object
      properties:
        channels:
          type: array
          items:
            $ref: '#/components/schemas/Channel'
          minItems: 1
          description: Notification channels to enable
        hours_before_expiry:
          type: integer
          minimum: 1
          description: Hours before expiry to send reminder
        frequency:
          $ref: '#/components/schemas/Frequency'
      required:
        - channels
        - hours_before_expiry
        - frequency
...
    ScenarioType:
      type: string
      enum:
        - no_check_ins
        - consistent_check_ins
        - missed_check_in_dates
      description: >
        The check-in behaviour scenario to simulate.

        - `no_check_ins` — owner never checks in again; vault releases when current TTL expires.
        - `consistent_check_ins` — owner checks in at every interval indefinitely; vault never releases.
        - `missed_check_in_dates` — owner misses N consecutive check-ins, then stops.
    ScenarioResult:
      type: object
      description: Projected release outcome for a single scenario
      properties:
        scenario:
          $ref: '#/components/schemas/ScenarioType'
        description:
          type: string
          description: Human-readable description of the scenario assumptions
        projected_release_at:
          type: string
          format: date-time
          description: >
            Projected UTC timestamp when the vault will release under this scenario.
            For `consistent_check_ins`, this is a far-future sentinel value.
        seconds_until_release:
          type: integer
          format: int64
          description: >
            Seconds from simulation time until projected release.
            -1 means the vault will never release under this scenario.
        confidence:
          type: string
          enum: [high, medium, low]
          description: >
            Confidence in the projection.
            `high` — deterministic given vault state.
            `medium` — 1–2 missed check-ins assumed.
            `low` — 3+ missed check-ins assumed (higher uncertainty).
        notes:
          type: string
          description: Additional context about the scenario's assumptions and calculations
      required:
        - scenario
        - description
        - projected_release_at
        - seconds_until_release
        - confidence
        - notes
    SimulateReleaseResponse:
      type: object
      description: Response containing simulation results for one or more release scenarios
      properties:
        vault_id:
          type: string
          description: The ID of the vault that was simulated
        current_ttl_remaining:
          type: integer
          format: int64
          nullable: true
          description: Current TTL remaining in seconds; null if not known or already expired
        check_in_interval:
          type: integer
          format: int64
          description: The vault's configured check-in interval in seconds
        last_check_in:
          type: string
          format: date-time
          description: Timestamp of the last recorded check-in
        scenarios:
          type: array
          items:
            $ref: '#/components/schemas/ScenarioResult'
          description: Simulation results, one per requested scenario
        simulated_at:
          type: string
          format: date-time
          description: Timestamp when this simulation was generated
      required:
        - vault_id
        - check_in_interval
        - last_check_in
        - scenarios
        - simulated_at

    # ── Legal Document Anchoring schemas (Issue #1339) ──────────────────────
    LegalDocumentType:
      type: string
      enum: [will, trust_deed, power_of_attorney, other]
      description: Category of the legal document being anchored
    AnchorDocumentRequest:
      type: object
      description: Request body for anchoring a legal document hash
      properties:
        owner_address:
          type: string
          description: Stellar account address of the vault owner
        doc_hash_hex:
          type: string
          pattern: '^[0-9a-fA-F]{64}$'
          description: Hex-encoded SHA-256 hash of the document (64 characters)
        doc_type:
          $ref: '#/components/schemas/LegalDocumentType'
        storage_ref:
          type: string
          maxLength: 128
          nullable: true
          description: Optional IPFS CID or off-chain storage reference (max 128 chars)
      required:
        - owner_address
        - doc_hash_hex
        - doc_type
    DocumentAnchorRecord:
      type: object
      description: An anchored legal document hash record
      properties:
        id:
          type: integer
          format: int64
          description: Database row ID
        vault_id:
          type: string
          description: Vault the anchor belongs to
        doc_id:
          type: integer
          format: int32
          description: Sequential document ID within the vault (1-indexed)
        doc_hash_hex:
          type: string
          pattern: '^[0-9a-fA-F]{64}$'
          description: Hex-encoded SHA-256 hash of the anchored document
        doc_type:
          $ref: '#/components/schemas/LegalDocumentType'
        storage_ref:
          type: string
          nullable: true
          description: Optional off-chain storage reference
        anchored_at:
          type: string
          format: date-time
          description: UTC timestamp when the anchor was recorded
        removed:
          type: boolean
          description: Whether the owner has soft-removed this anchor
      required:
        - id
        - vault_id
        - doc_id
        - doc_hash_hex
        - doc_type
        - anchored_at
        - removed
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token authentication for passkey-authenticated requests
security:
  - ApiKeyAuth: []
  - BearerAuth: []